+++ test/project_erlang_windows/test/erlang_test_file.erl
@@ -1,0 +1,9 @@
+-module(erlang_test_file).
+
+-export([main/0]).
+
+-include("erlang_test_header.hrl").
+
+main() ->
+  String = header_function(),
+  <<"Hello, from the Erlang test module!\n"/utf8, String/binary>>.