+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Imported modules can now be discarded by giving them an alias starting with `_`.
 
+### Build tool changes
+
+- Added Windows support for the `gleam export erlang-shipment` command.
+
 ## v0.32.2 - 2023-11-03
 
 ### Build tool changes